Updated and added some comments for the TM1829 controller
Please have a look at lines 56/57: Which oder is the TM1829? This information should also be mentioned there. As well as the information if the TM1829 is 800 or 400 KHz (lines 59-64).

PS: Great work! (and it even almost works for me without a level shifter... only sometimes a segment gets "stuck" and I have to connect the original controller for a few seconds, then it works with the spark again - so great for the first tests, but in the long term I need to get a level shifter... ;) )